Code refactoring + reorganization
