+++ index.html
@@ -5,7 +5,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script async src="js/main.js"></script>
   <title>Кексобукинг</title>
 </head>
 <body>
@@ -230,5 +229,6 @@
     </div>
   </template>
 
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
